refactor: unmerge server
