+++ auth/core/src/lib.rs
@@ -1,3 +1,2 @@
 pub use auth_api as api;
-pub use auth_server as server;
 pub use auth_types as types;
