--- auth/server/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "houseflow-auth-server"
-version = "0.1.0"
-authors = ["gbaranski <root@gbaranski.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-auth-types   = { path = "../types"   , package = "houseflow-auth-types", features = [ "actix", "db" ] }
-types        = { path = "../../types", package = "houseflow-types" }
-token        = { path = "../../token", package = "houseflow-token", features = [ "store", "actix" ] }
-db           = { path = "../../db"   , package = "houseflow-db" }
-config       = { path = "../../config", package = "houseflow-config", features = [ "server" ] }
-
-tokio        = { version = "1.6.0" }
-log          = "0.4.14"
-thiserror    = "1.0.25"
-rand         = "0.8.3"
-rust-argon2  = "0.8.3"
-serde_json   = "1.0.64"
-
-actix-web    = "4.0.0-beta.7" 
-actix-rt     = "2.2.0"
-validator    = "0.13.0"
-serde = { version = "1.0.126", features = ["derive"] }
-
-[dev-dependencies]
-hex = "0.4.3"
