--- auth/server/src/auth/mod.rs
@@ -1,3 +1,0 @@
-pub mod login;
-pub mod logout;
-pub mod register;
