+++ core/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+houseflow-server = { path = "../server" }
 types              = { path = "../types/"             , package = "houseflow-types" }
 token              = { path = "../token/"             , package = "houseflow-token", features = [ "store" ] }
 auth               = { path = "../auth/core/"         , package = "houseflow-auth"     }
