+++ core/src/server/run.rs
@@ -1,13 +1,13 @@
 use crate::{Command, ServerCommandState};
 use actix_web::{
-    web::{self, Data},
+    web::Data,
     App, HttpServer,
 };
 use anyhow::Context;
 use async_trait::async_trait;
 use db::Database;
 use std::sync::Arc;
-use token::store::TokenStore;
+use token::{store::TokenStore};
 
 use clap::Clap;
 
@@ -17,54 +17,25 @@ pub struct RunServerCommand {}
 #[async_trait(?Send)]
 impl Command<ServerCommandState> for RunServerCommand {
     async fn run(&self, state: ServerCommandState) -> anyhow::Result<()> {
-        let token_store = || async {
-            token::store::RedisTokenStore::new()
-                .await
-                .with_context(|| "connect to redis failed, is redis on?")
-        };
+        let token_store = token::store::RedisTokenStore::new()
+            .await
+            .with_context(|| "connect to redis failed, is redis on?")?;
+        let token_store = Data::from(Arc::new(token_store) as Arc<dyn TokenStore>);
 
-        let database = || async {
-            db::postgres::Database::new(&state.config.postgres)
-                .await
-                .with_context(|| "connect to postgres failed, is postgres on?")
-        };
-
-        let lighthouse_api = || lighthouse::api::Lighthouse::new(state.config.address);
-
-        let database = Data::from(Arc::from(database().await?) as Arc<dyn Database>);
-        let token_store = Data::from(Arc::from(token_store().await?) as Arc<dyn TokenStore>);
-        let lighthouse_api = Data::from(
-            Arc::from(lighthouse_api()) as Arc<dyn lighthouse::api::prelude::Lighthouse>
-        );
-        let lighthouse_app_data = Data::from(Arc::from(lighthouse::server::AppState::default()));
+        let database = db::postgres::Database::new(&state.config.postgres)
+            .await
+            .with_context(|| "connect to postgres failed, is postgres on?")?;
+        let database = Data::from(Arc::new(database) as Arc<dyn Database>);
 
         let address = state.config.address;
+        let secrets = Data::new(state.config.secrets);
+        let sessions = Data::new(houseflow_server::Sessions::default());
         let server = HttpServer::new(move || {
             App::new()
                 .wrap(actix_web::middleware::Logger::default())
-                .service(web::scope("/auth").configure(|cfg| {
-                    auth::server::configure(
-                        cfg,
-                        token_store.clone(),
-                        database.clone(),
-                        state.config.secrets.clone(),
-                    )
-                }))
-                .service(web::scope("/lighthouse").configure(|cfg| {
-                    lighthouse::server::configure(
-                        cfg,
-                        lighthouse_app_data.clone(),
-                        database.clone(),
-                    )
-                }))
-                .service(web::scope("/fulfillment").configure(|cfg| {
-                    fulfillment::server::configure(
-                        cfg,
-                        database.clone(),
-                        lighthouse_api.clone(),
-                        state.config.secrets.clone(),
-                    )
-                }))
+                .configure(|cfg| {
+                    houseflow_server::configure(cfg, token_store.clone(), database.clone(), secrets.clone(), sessions.clone())
+                })
         })
         .bind(address)?;
         server.run().await?;
