+++ fulfillment/core/Cargo.toml
@@ -7,6 +7,5 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-fulfillment-server = { path = "../server/", package = "houseflow-fulfillment-server" }
 fulfillment-types  = { path = "../types/" , package = "houseflow-fulfillment-types" }
 fulfillment-api    = { path = "../api/"   , package = "houseflow-fulfillment-api" }
