+++ fulfillment/core/src/lib.rs
@@ -1,3 +1,2 @@
 pub use fulfillment_api as api;
-pub use fulfillment_server as server;
 pub use fulfillment_types as types;
