--- fulfillment/server/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "houseflow-fulfillment-server"
-version = "0.1.0"
-authors = ["gbaranski <root@gbaranski.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-db                = { path = "../../db"   	   , package = "houseflow-db" }
-types             = { path = "../../types"	   , package = "houseflow-types" }
-fulfillment-types = { path = "../types"   	   , package = "houseflow-fulfillment-types", features = [ "actix", "db" ] }
-lighthouse-api    = { path = "../../lighthouse/api", package = "houseflow-lighthouse-api" }
-lighthouse-proto  = { path = "../../lighthouse/proto", package = "houseflow-lighthouse-proto" }
-token             = { path = "../../token"         , package = "houseflow-token", features = [ "actix" ] }
-config       	  = { path = "../../config"        , package = "houseflow-config", features = [ "server" ] }
-
-
-tokio 	          = { version = "1.6.1"  , features = ["macros"] }
-serde 		  = { version = "1.0.126", features = ["derive"] }
-actix-web         = "4.0.0-beta.7" 
-log               = "0.4"
-thiserror         = "1.0"
-actix-rt 	  = "2.2.0"
-rand 	  	  = "0.8.3"
-semver 		  = "1.0.3"
-
-[dev-dependencies]
-futures = "0.3.15"
-hex = "0.4.3"
-rand = "0.8.3"
-semver = "1.0.3"
