--- lighthouse/api/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "houseflow-lighthouse-api"
-version = "0.1.0"
-authors = ["gbaranski <root@gbaranski.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-types 		 = { path = "../../types", package = "houseflow-types" }
-lighthouse-proto = { path = "../proto"   , package = "houseflow-lighthouse-proto" }
-lighthouse-types = { path = "../types"   , package = "houseflow-lighthouse-types" }
-
-reqwest 	 = { version = "0.11.4",     features = [ "json" ] }
-serde 		 =   "1.0"
-serde_json 	 =   "1.0"
-thiserror 	 =   "1.0"
-url 		 =   "2.2"
-async-trait = "0.1.50"
-tokio = { version = "1.6.1", features = ["sync"] }
-
