+++ lighthouse/core/src/lib.rs
@@ -1,4 +1,2 @@
-pub use lighthouse_api as api;
 pub use lighthouse_proto as proto;
-pub use lighthouse_server as server;
 pub use lighthouse_types as types;
