--- lighthouse/server/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "houseflow-lighthouse-server"
-version = "0.1.0"
-authors = ["gbaranski <root@gbaranski.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lighthouse-proto = { path = "../proto"   , package = "houseflow-lighthouse-proto" }
-lighthouse-types = { path = "../types"   , package = "houseflow-lighthouse-types", features = ["actix"]}
-db               = { path = "../../db"   	   , package = "houseflow-db" }
-types  		 = { path = "../../types", package = "houseflow-types" }
-
-actix 	 	 = "0.11.1"
-actix-web 	 = "4.0.0-beta.7" 
-actix-web-actors = "4.0.0-beta.5"
-actix-service 	 = "2.0.0-beta.5"
-tokio            = { version = "1.5", features = ["sync" ] }
-
-thiserror        = "1.0"
-log              = "0.4"
-
-itertools        = "0.10"
-
-bytes            = "1.0"
-serde            = "1.0"
-serde_json       = "1.0"
-rust-argon2 = "0.8.3"
