+++ server/src/lighthouse/mod.rs
@@ -1,0 +1,6 @@
+mod session;
+pub(crate) mod aliases;
+mod connect;
+
+pub use session::Session;
+pub use connect::on_websocket;
